complited first task
