+++ lab_1/first_task/key.json
@@ -1,0 +1,66 @@
+{
+	"а" : "е",
+	"б" : "ж",
+	"в" : "з",
+	"г" : "и",
+	"д" : "й",
+	"е" : "к",
+	"ж" : "л",
+	"з" : "м",
+	"и" : "н",
+	"й" : "о",
+	"к" : "п",
+	"л" : "р",
+	"м" : "с",
+	"н" : "т",
+	"о" : "у",
+	"п" : "ф",
+	"р" : "х",
+	"с" : "ц",
+	"т" : "ч",
+	"у" : "ш",
+	"ф" : "щ",
+	"х" : "ъ",
+	"ц" : "ы",
+	"ч" : "ь",
+	"ш" : "э",
+	"щ" : "ю",
+	"ъ" : "я",
+	"ы" : "а",
+	"ь" : "б",
+	"э" : "в",
+	"ю" : "г",
+	"я" : "д",
+	"А" : "Е",
+	"Б" : "Ж",
+	"В" : "З",
+	"Г" : "И",
+	"Д" : "Й",
+	"Е" : "К",
+	"Ж" : "Л",
+	"З" : "М",
+	"И" : "Н",
+	"Й" : "О",
+	"К" : "П",
+	"Л" : "Р",
+	"М" : "С",
+	"Н" : "Т",
+	"О" : "У",
+	"П" : "Ф",
+	"Р" : "Х",
+	"С" : "Ц",
+	"Т" : "Ч",
+	"У" : "Ш",
+	"Ф" : "Щ",
+	"Х" : "Ъ",
+	"Ц" : "Ы",
+	"Ч" : "Ь",
+	"Ш" : "Э",
+	"Щ" : "Ю",
+	"Ъ" : "Я",
+	"Ы" : "А",
+	"Ь" : "Б",
+	"Э" : "В",
+	"Ю" : "Г",
+	"Я" : "Д"
+}
