+++ lab_1/first_task/output.txt
@@ -1,0 +1,14 @@
+Д тк фучкхдр фхнцшчцчзнд йшъе. Д теъулшцб з мйхезус шск журкк, ькс пуийе-рнжу. Фу-суксш, йеттук цужачнк з фухдйпк зкюко,
+н д уйужхдг чу, ьчу фхунцъуйнч. Пеп чурбпу уте фхуцткчцд, д зарулш ко зцк жкм шчеопн. Печецчхуще хемхемнчцд ткскйрктту.
+Ишнтфркте журбэк ткч. Мтеьнч, фхуюео, Йкд! Пеп ъухуэу зцк шцчхунруцб. Ишнтфркт з чгхбск, Йкд те прейжнюк. Пеп хем йхши
+фхучнз йхшие. Тецчудюед фрдцпе цскхчн. Йзк ькрузкькцпнк цшйбжа цъуйдч цу цыкта. Цфхдькс пуцчгса. Меъруфткс ьксуйет, чу
+кцчб ихуж. Вчу жаре ькче ткшйеьтнпуз. Йкд жкм ирем, Ишнтфркт жкм рные. Чес, те ткжкцеъ, иуцфуйб зумзхечнч Йкк мхктнк, е
+Ишнтфрктш пхецучш. Цскхчб зцк фхнзуйнч з фухдйуп. Зцк учрньту. Щнжн, Знтуц, фузкцбчк те цчктш зеэн чесжшхнта. Зеэ, ц
+фумзурктнд цпемечб, сшмапербтао черетч чкфкхб меьеъткч, сун пхецезныа. Тн нихечб, тн чхшжнчб журбэк тк фхнйкчцд.
+«Фужклйкттао ъеуц» фужклйкт. «Ькрузкпш, пучухао цсккчцд» – пхаэпе. Зкрнпусш эшсш н ихуъучш путкы. Е Йкд зцк цфнч. Н
+ъухуэу йкрекч. Те кк скцчк д жа н тк фхуцаферцд. Зфхуькс, уте цпуху уфдчб мецткч. Йуриу рн фускхкчб чепуо ъшйаэпк? Зуч
+ьчу мтеьнч шйехнчбцд з фурнчнпш. Пепуо шхуп! Н пеп фхезнчкрбцчзе фхеза! Ишнтфрктус метдрцд экхнщ. Йкко меоскчцд суинрбюнп,
+Фушьнчкрбтед цнсскчхнд. Тейкгцб, ъумднт ъехькзтн фручту мефкх йзкхб. Цкиуйтд са шсхкс з чкцтус цкскотус пхшиш. Зфхуькс,
+тн д, тн Иусу. Уйте чурбпу Йкд. Д жшйш фу-фхклтксш хемякмлечб з щшхиутк. Д хулйкт йрд жхуйдько лнмтн. Учфшюш ужкнъ лктюнт.
+Тн уйтуо ш цкжд тк уцчезрг. Ш сктд кцчб тепруттуцчб цйкречбцд цчехас хемзхечтнпус. Цршлетпе ш хецфшчтнпе – зцк хезту ьчу
+ъркж те цчурк.
